fix: ensure absolute OG Image URLs using metadataBase and enable sharing on history routes
